pyproject.toml
[project]
name = "grok-trader"
version = "0.1.0"
description = "Dominate markets with data."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "praw>=7.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tweepy>=4.16.0",
    "xai-sdk>=1.5.0",
    "websockets>=12.0",
    "bs4>=0.0.2",
    "polymarket>=0.1.2",
    "websocket-client>=1.9.0",
]
